+++ build.sh
@@ -11,6 +11,7 @@ src/Compose.js
 src/Camera.js
 src/RenderTarget.js
 src/ScreenGeometry.js
+src/ShaderMaterial.js
 src/FragmentMaterial.js
 src/DownsampleMaterial.js
 src/RaytraceMaterial.js
@@ -35,22 +36,22 @@ cat $SHADERS > build/ThreeRTT.glsl.html
 
 if [ -z "$SKIP_MINIFY" ]; then
 curl --data-urlencode "js_code@build/ThreeRTT.js" 	\
-	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS" \
+	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS&language=ECMASCRIPT5" \
 	http://closure-compiler.appspot.com/compile	\
 	> build/ThreeRTT.min.js
 
 curl --data-urlencode "js_code@build/ThreeRTT-tquery.js" 	\
-	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS" \
+	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS&language=ECMASCRIPT5" \
 	http://closure-compiler.appspot.com/compile	\
 	> build/ThreeRTT-tquery.min.js
 
 curl --data-urlencode "js_code@build/ThreeRTT-core.js" 	\
-	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS" \
+	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS&language=ECMASCRIPT5" \
 	http://closure-compiler.appspot.com/compile	\
 	> build/ThreeRTT-core.min.js
 
 curl --data-urlencode "js_code@build/ThreeRTT-core-tquery.js" 	\
-	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS" \
+	-d "output_format=text&output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS&language=ECMASCRIPT5" \
 	http://closure-compiler.appspot.com/compile	\
 	> build/ThreeRTT-core-tquery.min.js
 fi