+++ src/tQuery/tQuery.js
@@ -9,7 +9,17 @@ tQuery.World.registerInstance('rtt', function (options) {
  * Add a surface showing a render-to-texture surface to this world.
  */
 tQuery.World.registerInstance('compose', function (rtts, fragmentShader, textures, uniforms) {
-  tQuery.composeRTT(this, rtts, fragmentShader, textures, uniforms);
+  var compose = tQuery.createComposeRTT(rtts, fragmentShader, textures, uniforms);
+  this.add(compose);
+  return this;
+});
+
+/**
+ * Add a surface showing a render-to-texture surface to this world.
+ */
+tQuery.World.registerInstance('display', function (targets, gx, gy) {
+  var display = tQuery.createDisplayRTT(targets, gx, gy);
+  this.add(display);
   return this;
 });
 
@@ -22,10 +32,24 @@ tQuery.registerStatic('createRTT', function (world, options) {
 });
 
 /**
- * Create a surface showing a render-to-texture image in this world.
+ * Composite a render-to-texture image full screen
  */
-tQuery.registerStatic('composeRTT', function (world, rtts, fragmentShader, textures, uniforms) {
-  return new ThreeRTT.Compose(world.tScene(), rtts, fragmentShader, textures, uniforms);
+tQuery.registerStatic('createComposeRTT', function (rtts, fragmentShader, textures, uniforms) {
+  return new ThreeRTT.Compose(rtts, fragmentShader, textures, uniforms);
+});
+
+/**
+ * Create a display surface showing one or more textures in a grid.
+ */
+tQuery.registerStatic('createDisplayRTT', function (targets, gx, gy) {
+  return new ThreeRTT.Display(targets, gx, gy);
+});
+
+/**
+ * Create a ShaderMaterial.
+ */
+tQuery.registerStatic('createShaderMaterial', function (worlds, vertexShader, fragmentShader, textures, uniforms) {
+  return new ThreeRTT.FragmentMaterial(worlds, vertexShader, fragmentShader, textures, uniforms);
 });
 
 /**